# openapi.json
{"openapi":"3.1.0","info":{"title":"video generator api","version":"0.1.0"},"paths":{"/generate":{"post":{"summary":"Generate Video","operationId":"generate_video_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VideoRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VideoResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/get_id_for_script":{"post":{"summary":"Get Id For Script","operationId":"get_id_for_script_get_id_for_script_post","parameters":[{"name":"script","in":"query","required":true,"schema":{"type":"string","title":"Script"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/status/{video_id}":{"get":{"summary":"Get Video Status","operationId":"get_video_status_status__video_id__get","parameters":[{"name":"video_id","in":"path","required":true,"schema":{"type":"string","title":"Video Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VideoResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/status":{"get":{"summary":"Get All Videos Status","operationId":"get_all_videos_status_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/download/{video_id}":{"get":{"summary":"Download Video","operationId":"download_video_download__video_id__get","parameters":[{"name":"video_id","in":"path","required":true,"schema":{"type":"string","title":"Video Id"}}],"responses":{"200":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VideoRequest":{"properties":{"script":{"type":"string","title":"Script"},"topic":{"type":"string","title":"Topic","default":"technology"},"portrait":{"type":"boolean","title":"Portrait","default":true}},"type":"object","required":["script"],"title":"VideoRequest"},"VideoResponse":{"properties":{"video_id":{"type":"string","title":"Video Id"},"state":{"$ref":"#/components/schemas/VideoState"}},"type":"object","required":["video_id","state"],"title":"VideoResponse"},"VideoState":{"type":"string","enum":["processing","done","error"],"title":"VideoState"}}}}